Implement Clone trait on structures + support multiple events
Added the `Clone` trait to a bunch of structures so I wasn't having to borrow references in the tests.

Added support for new `Individual` fields, including making `birth` and `death` vectors b/c it's possible for a GEDCOM to contain duplicate events. AFAICT, the first event type encountered should be considered the `preferred` event.