Fix parsing of ADOP records
There was a discrepency in parsing because I'd copy/pasted the test for
an Adoption, which contained test data for an Individual Record, but not
an Adoption record.

I added some logic to make sure that we're parsing an ADOP record and
cleaned up the test to check a subset of the event details.